requirements-dev.txt
# requirements-dev.txt (para desarrollo)
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
factory-boy>=3.3.0  # Para crear datos de prueba
django-debug-toolbar>=4.2.0  # Para debugging